requirements.txt
Django==2.0.7
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
imutils==0.4.6
numpy==1.15.0
opencv-python==3.4.2.17
scipy==1.1.0
django-multiselectfield==0.1.8
